refactor: split ScheduledJobForm and ScheduledJobsView into focused modules
Extract the two oversized scheduled-jobs components into cohesive pieces so a
typical change touches one file:

- ScheduledJobsView (731 -> 303): now data/orchestration only; rendering moved
  to JobsList (list view, with the row actions menu deduped) and JobRunDetail
  (run-detail view), with shared display helpers in helpers.tsx.
- ScheduledJobForm (991 -> 464): now pure layout; all state/effects/handlers
  moved to lib/hooks/useScheduledJobForm.ts. Constants and timezone/interval
  helpers moved to form-config.ts, and the schedule picker to ScheduleFields.

No behavior change.
